+++ CHANGES.rst
@@ -4,6 +4,9 @@ Histórico de Alterações
 1.1.2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+* Corrige conteúdo inicial para novas versões do collective.cover. (closes `#216`_).
+  [rodfersou]
+
 * Todas as pastas da raiz precisam ter a ordenação padrão do Plone. (closes `#190`_).
   [idgserpro]
 
@@ -379,3 +382,4 @@ Histórico de Alterações
 .. _`#175`: https://github.com/plonegovbr/brasil.gov.portal/issues/175
 .. _`#182`: https://github.com/plonegovbr/brasil.gov.portal/issues/182
 .. _`#190`: https://github.com/plonegovbr/brasil.gov.portal/issues/190
+.. _`#216`: https://github.com/plonegovbr/brasil.gov.portal/issues/216